+++ packages/calcite-components/src/components/inline-editable/inline-editable.e2e.ts
@@ -1,9 +1,10 @@
 // @ts-strict-ignore
-import { newE2EPage, E2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
-import { describe, expect, it, beforeEach } from "vitest";
-import { accessible, disabled, labelable, renders, hidden, t9n } from "../../tests/commonTests";
+import { E2EPage, newE2EPage } from "@arcgis/lumina-compiler/puppeteerTesting";
+import { beforeEach, describe, expect, it } from "vitest";
+import { accessible, disabled, hidden, labelable, renders, t9n, themed } from "../../tests/commonTests";
 import { html } from "../../../support/formatting";
 import type { Input } from "../input/input";
+import { findAll } from "../../tests/utils";
 import { CSS } from "./resources";
 import type { InlineEditable } from "./inline-editable";
 
@@ -56,7 +57,7 @@ describe("calcite-inline-editable", () => {
         </calcite-inline-editable>`,
       });
 
-      const buttons = await page.findAll("calcite-inline-editable >>> calcite-button");
+      const buttons = await findAll(page, "calcite-inline-editable >>> calcite-button");
 
       expect(buttons).toHaveLength(3);
 
@@ -72,7 +73,7 @@ describe("calcite-inline-editable", () => {
         </calcite-inline-editable>`,
       });
 
-      const buttons = await page.findAll("calcite-inline-editable >>> calcite-button");
+      const buttons = await findAll(page, "calcite-inline-editable >>> calcite-button");
 
       expect(buttons).toHaveLength(3);
 
@@ -398,5 +399,72 @@ describe("calcite-inline-editable", () => {
 
   describe("translation support", () => {
     t9n("calcite-inline-editable");
+  });
+
+  describe("theme", () => {
+    themed("calcite-inline-editable", {
+      "--calcite-inline-editable-background-color-hover": {
+        shadowSelector: `.${CSS.wrapper}`,
+        state: "hover",
+        targetProp: "backgroundColor",
+      },
+      "--calcite-inline-editable-background-color": {
+        shadowSelector: `.${CSS.wrapper}`,
+        targetProp: "backgroundColor",
+      },
+    });
+    themed(
+      html`<calcite-inline-editable controls editing-enabled>
+        <calcite-input />
+      </calcite-inline-editable>`,
+      {
+        "--calcite-inline-editable-button-corner-radius": [
+          {
+            shadowSelector: `.${CSS.enableEditingButton}`,
+            targetProp: "--calcite-button-corner-radius",
+          },
+          {
+            shadowSelector: `.${CSS.cancelEditingButton}`,
+            targetProp: "--calcite-button-corner-radius",
+          },
+          {
+            shadowSelector: `.${CSS.confirmChangesButton}`,
+            targetProp: "--calcite-button-corner-radius",
+          },
+        ],
+        "--calcite-inline-editable-button-loader-color": {
+          shadowSelector: `.${CSS.confirmChangesButton}`,
+          targetProp: "--calcite-button-loader-color",
+        },
+        "--calcite-inline-editable-button-shadow-color": [
+          {
+            shadowSelector: `.${CSS.enableEditingButton}`,
+            targetProp: "--calcite-button-shadow-color",
+          },
+          {
+            shadowSelector: `.${CSS.cancelEditingButton}`,
+            targetProp: "--calcite-button-shadow-color",
+          },
+          {
+            shadowSelector: `.${CSS.confirmChangesButton}`,
+            targetProp: "--calcite-button-shadow-color",
+          },
+        ],
+        "--calcite-inline-editable-button-text-color": [
+          {
+            shadowSelector: `.${CSS.enableEditingButton}`,
+            targetProp: "--calcite-button-text-color",
+          },
+          {
+            shadowSelector: `.${CSS.cancelEditingButton}`,
+            targetProp: "--calcite-button-text-color",
+          },
+          {
+            shadowSelector: `.${CSS.confirmChangesButton}`,
+            targetProp: "--calcite-button-text-color",
+          },
+        ],
+      },
+    );
   });
 });